Autenticação das Views

+++ SoftCandy/Controllers/ClienteController.cs
@@ -1,4 +1,4 @@
-﻿ using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -23,31 +23,43 @@ namespace SoftCandy.Controllers
         // GET: Cliente
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Cliente.ToListAsync());
+            if (User.Identity.IsAuthenticated)
+            {
+                return View(await _context.Cliente.ToListAsync());
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cliente/Details
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Id não foi fornecido!" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não foi fornecido!" });
+                }
 
-            var cliente = await _context.Cliente
-                .FirstOrDefaultAsync(m => m.Id_Cliente == id);
-            if (cliente == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
-            }
+                var cliente = await _context.Cliente
+                    .FirstOrDefaultAsync(m => m.Id_Cliente == id);
+                if (cliente == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
+                }
 
-            return View(cliente);
+                return View(cliente);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cliente/Create
         public IActionResult Create()
         {
-            return View();
+            if (User.Identity.IsAuthenticated)
+            {
+                return View();
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cliente/Create
@@ -55,29 +67,37 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("id_Cliente,Nome,Celular,Endereco")] Cliente cliente)
         {
-            if (ModelState.IsValid)
+            if (User.Identity.IsAuthenticated)
             {
-                _context.Add(cliente);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(cliente);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(cliente);
             }
-            return View(cliente);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cliente/Edit
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Id não fornecido!" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não fornecido!" });
+                }
 
-            var cliente = await _context.Cliente.FindAsync(id);
-            if (cliente == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
+                var cliente = await _context.Cliente.FindAsync(id);
+                if (cliente == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
+                }
+                return View(cliente);
             }
-            return View(cliente);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cliente/Edit
@@ -85,50 +105,58 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id_Cliente,Nome,Celular,Endereco")] Cliente cliente)
         {
-            if (id != cliente.Id_Cliente)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Id não corresponde!" });
-            }
+                if (id != cliente.Id_Cliente)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não corresponde!" });
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(cliente);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException e)
-                {
-                    if (!ClienteExists(cliente.Id_Cliente))
+                    try
                     {
-                        return RedirectToAction(nameof(Error), new { message = e.Message });
+                        _context.Update(cliente);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException e)
                     {
-                        throw;
+                        if (!ClienteExists(cliente.Id_Cliente))
+                        {
+                            return RedirectToAction(nameof(Error), new { message = e.Message });
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                return View(cliente);
             }
-            return View(cliente);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cliente/Delete
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Id não foi fornecido!" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não foi fornecido!" });
+                }
 
-            var cliente = await _context.Cliente
-                .FirstOrDefaultAsync(m => m.Id_Cliente == id);
-            if (cliente == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
-            }
+                var cliente = await _context.Cliente
+                    .FirstOrDefaultAsync(m => m.Id_Cliente == id);
+                if (cliente == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Id não existe!" });
+                }
 
-            return View(cliente);
+                return View(cliente);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cliente/Delete
@@ -136,25 +164,33 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var cliente = await _context.Cliente.FindAsync(id);
-            _context.Cliente.Remove(cliente);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (User.Identity.IsAuthenticated)
+            {
+                var cliente = await _context.Cliente.FindAsync(id);
+                _context.Cliente.Remove(cliente);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool ClienteExists(int id)
         {
             return _context.Cliente.Any(e => e.Id_Cliente == id);
         }
-       
+
         public IActionResult Error(string message)
         {
-            var viewModel = new ErrorViewModel
+            if (User.Identity.IsAuthenticated)
             {
-               Message = message,
-               RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
-            };
-            return View(viewModel);
+                var viewModel = new ErrorViewModel
+                {
+                    Message = message,
+                    RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+                };
+                return View(viewModel);
+            }
+            return RedirectToAction("Index", "Home");
         }
     }
 }
