+++ SoftCandy/Controllers/Item_PedidoController.cs
@@ -22,123 +22,147 @@ namespace SoftCandy.Controllers
         // GET: Item_Pedido
         public async Task<IActionResult> Index()
         {
-            var softCandyContext = _context.Item_Pedido.Include(i => i.Produto);
-            return View(await softCandyContext.ToListAsync());
+            if (User.Identity.IsAuthenticated)
+            {
+                var softCandyContext = _context.Item_Pedido.Include(i => i.Produto);
+                return View(await softCandyContext.ToListAsync());
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Item_Pedido/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var item_Pedido = await _context.Item_Pedido
-                .Include(i => i.Produto)
-                .FirstOrDefaultAsync(m => m.Num_Pedido == id);
-            if (item_Pedido == null)
-            {
-                return NotFound();
-            }
+                var item_Pedido = await _context.Item_Pedido
+                    .Include(i => i.Produto)
+                    .FirstOrDefaultAsync(m => m.Num_Pedido == id);
+                if (item_Pedido == null)
+                {
+                    return NotFound();
+                }
 
-            return View(item_Pedido);
+                return View(item_Pedido);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Item_Pedido/Create
         public IActionResult Create()
         {
-            ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto");
-            return View();
+            if (User.Identity.IsAuthenticated)
+            {
+                ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto");
+                return View();
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Item_Pedido/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Preco_Pago,Quantidade,Cod_Produto,Num_Pedido")] Item_Pedido item_Pedido)
         {
-            if (ModelState.IsValid)
+            if (User.Identity.IsAuthenticated)
             {
-                _context.Add(item_Pedido);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(item_Pedido);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
+                return View(item_Pedido);
             }
-            ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
-            return View(item_Pedido);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Item_Pedido/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var item_Pedido = await _context.Item_Pedido.FindAsync(id);
-            if (item_Pedido == null)
-            {
-                return NotFound();
+                var item_Pedido = await _context.Item_Pedido.FindAsync(id);
+                if (item_Pedido == null)
+                {
+                    return NotFound();
+                }
+                ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
+                return View(item_Pedido);
             }
-            ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
-            return View(item_Pedido);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Item_Pedido/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Preco_Pago,Quantidade,Cod_Produto,Num_Pedido")] Item_Pedido item_Pedido)
         {
-            if (id != item_Pedido.Num_Pedido)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id != item_Pedido.Num_Pedido)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(item_Pedido);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!Item_PedidoExists(item_Pedido.Num_Pedido))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(item_Pedido);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!Item_PedidoExists(item_Pedido.Num_Pedido))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
+                return View(item_Pedido);
             }
-            ViewData["Cod_Produto"] = new SelectList(_context.Produto, "Cod_Produto", "Nome_Produto", item_Pedido.Cod_Produto);
-            return View(item_Pedido);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Item_Pedido/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var item_Pedido = await _context.Item_Pedido
-                .Include(i => i.Produto)
-                .FirstOrDefaultAsync(m => m.Num_Pedido == id);
-            if (item_Pedido == null)
-            {
-                return NotFound();
-            }
+                var item_Pedido = await _context.Item_Pedido
+                    .Include(i => i.Produto)
+                    .FirstOrDefaultAsync(m => m.Num_Pedido == id);
+                if (item_Pedido == null)
+                {
+                    return NotFound();
+                }
 
-            return View(item_Pedido);
+                return View(item_Pedido);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Item_Pedido/Delete/5
@@ -146,10 +170,14 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var item_Pedido = await _context.Item_Pedido.FindAsync(id);
-            _context.Item_Pedido.Remove(item_Pedido);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (User.Identity.IsAuthenticated)
+            {
+                var item_Pedido = await _context.Item_Pedido.FindAsync(id);
+                _context.Item_Pedido.Remove(item_Pedido);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool Item_PedidoExists(int id)
