+++ SoftCandy/Controllers/ProdutoController.cs
@@ -23,31 +23,43 @@ namespace SoftCandy.Controllers
         // GET: Produto
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Produto.ToListAsync());
+            if (User.Identity.IsAuthenticated)
+            {
+                return View(await _context.Produto.ToListAsync());
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Produto/Details
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido" });
+                }
 
-            var produto = await _context.Produto
-                .FirstOrDefaultAsync(m => m.Cod_Produto == id);
-            if (produto == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
-            }
+                var produto = await _context.Produto
+                    .FirstOrDefaultAsync(m => m.Cod_Produto == id);
+                if (produto == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
+                }
 
-            return View(produto);
+                return View(produto);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Produto/Create
         public IActionResult Create()
         {
-            return View();
+            if (User.Identity.IsAuthenticated)
+            {
+                return View();
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Produto/Create
@@ -55,29 +67,37 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Cod_Produto,Nome_Produto,Quantidade,Preco_Venda,Descricao")] Produto produto)
         {
-            if (ModelState.IsValid)
+            if (User.Identity.IsAuthenticated)
             {
-                _context.Add(produto);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(produto);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(produto);
             }
-            return View(produto);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Produto/Edit
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido" });
+                }
 
-            var produto = await _context.Produto.FirstOrDefaultAsync(m=>m.Cod_Produto==id);
-            if (produto == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
+                var produto = await _context.Produto.FirstOrDefaultAsync(m => m.Cod_Produto == id);
+                if (produto == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
+                }
+                return View(produto);
             }
-            return View(produto);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Produto/Edit
@@ -85,50 +105,59 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Cod_Produto,Nome_Produto,Quantidade,Preco_Venda,Descricao")] Produto produto)
         {
-            if (id != produto.Cod_Produto)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Código não corresponde!" });
-            }
+                if (id != produto.Cod_Produto)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não corresponde!" });
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(produto);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException e)
-                {
-                    if (!ProdutoExists(produto.Cod_Produto))
+                    try
                     {
-                        return RedirectToAction(nameof(Error), new { message = e.Message });
+                        _context.Update(produto);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException e)
                     {
-                        throw;
+                        if (!ProdutoExists(produto.Cod_Produto))
+                        {
+                            return RedirectToAction(nameof(Error), new { message = e.Message });
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                return View(produto);
             }
-            return View(produto);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Produto/Delete
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido!" });
-            }
+                if (id == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não foi fornecido!" });
 
-            var produto = await _context.Produto
-                .FirstOrDefaultAsync(m => m.Cod_Produto == id);
-            if (produto == null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
-            }
+                }
 
-            return View(produto);
+                var produto = await _context.Produto
+                    .FirstOrDefaultAsync(m => m.Cod_Produto == id);
+                if (produto == null)
+                {
+                    return RedirectToAction(nameof(Error), new { message = "Código não existe!" });
+                }
+
+                return View(produto);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Produto/Delete
@@ -136,10 +165,14 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var produto = await _context.Produto.FindAsync(id);
-            _context.Produto.Remove(produto);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (User.Identity.IsAuthenticated)
+            {
+                var produto = await _context.Produto.FindAsync(id);
+                _context.Produto.Remove(produto);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool ProdutoExists(int id)
@@ -148,13 +181,16 @@ namespace SoftCandy.Controllers
         }
         public IActionResult Error(string message)
         {
-            var viewModel = new ErrorViewModel
+            if (User.Identity.IsAuthenticated)
             {
-                Message = message,
-                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
-            };
-            return View(viewModel);
-
+                var viewModel = new ErrorViewModel
+                {
+                    Message = message,
+                    RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+                };
+                return View(viewModel);
+            }
+            return RedirectToAction("Index", "Home");
 
         }
     }
